Move the downward API directory to '/etc/podnetinfo'
app-netutil move the downward API directory to '/etc/podnetinfo' to
match SR-IOV admission controller value. Make the same change to this
repo.

Also moved all the podspecs back to using the latest locally built
image. This was locked to a fixed version DockerHub version in a
previous merge to lock down images for KubeCon Demo.

Signed-off-by: Billy McFall <22157057+Billy99@users.noreply.github.com>